+++ app/Config/Routes.php
@@ -93,7 +93,7 @@ $routes->group('', ['filter' => ['auth', 'admin-area']], function (RouteCollecti
         $routes->get('(:num)/edit',     'Interventi::edit/$1');
         $routes->post('(:num)',         'Interventi::update/$1');
         $routes->post('(:num)/elimina', 'Interventi::delete/$1');
-        $routes->get('(:num)/chiudi',    'Interventi::chiudi/$1');
+        $routes->post('(:num)/chiudi',   'Interventi::chiudi/$1');
         $routes->post('(:num)/assegna', 'Interventi::assegnaTecnico/$1');
         $routes->get('(:num)/pdf',      'Interventi::pdf/$1');
     });